requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==3.2.16
django-js-asset==2.2.0
django-mptt==0.16.0
django-mptt-admin==2.7.0
flake8==7.1.1
isort==5.13.2
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.1
mccabe==0.7.0
mypy==1.12.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
template==0.7.6
toml==0.10.2
tomli==2.0.2
typing_extensions==4.12.2
